updated file location of enum values
